groundbaddiepush: retype groundBaddiePushPlayerOut/guardClawUpdateWhileFrozen obj int->GameObject*, drop 8 self-casts; seqobj11e/Baddie.c call sites updated (all 3 .o byte-identical, ninja+DOL OK)

+++ src/main/dll/seqobj11e.c
@@ -148,7 +148,7 @@ void guardClaw_update(GameObject* obj, u8* state)
     }
     if (gSeq11EStateTable[((BaddieState*)state)->userData1].flagB != 0)
     {
-        groundBaddiePushPlayerOut((int)obj, state);
+        groundBaddiePushPlayerOut(obj, state);
     }
 }
